added/changed issues: * moved common code scopes to the Commons.scala * fixed inheritance test running
